requirements.txt

streamlit==1.39.0
openai>=1.53.0
fpdf2>=2.7.8
Pillow>=10.4.0
